% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmediate.R
\docType{data}
\name{sim}
\alias{sim}
\title{A simulated example data}
\value{
A list containing
\item{site}{Site ID}
\item{y}{Outcome}
\item{tr}{Treatment}
\item{me}{Mediator}
\item{x1}{Pretreatment covariate}
\item{x2}{Pretreatment covariate}
\item{x3}{Pretreatment covariate}
\item{x4}{Pretreatment covariate}
}
\description{
This simulated data list is for demonstration.
}
