% Please edit documentation in R/msmediate.R
\docType{data}
\name{sim.weights}
\alias{sim.weights}
\title{A simulated example data}
\value{
A list containing
\item{site}{Site ID}
\item{y}{Outcome}
\item{tr}{Treatment}
\item{me}{Mediator}
\item{x1}{Pretreatment covariate}
\item{x2}{Pretreatment covariate}
\item{x3}{Pretreatment covariate}
\item{R}{Response indicator}
\item{WD}{Sample weight}
}
\description{
This simulated data list is for demonstration.
}
